Userdata folder.
Now instead of storing data in the application folder,
it stores it in the user data folder.
This prevents data loss in case the folder with the program is deleted.
A seed database is included in the application and it's copied to
the userdata folder when the program is first launched.
The program always loads the same database so it can be updated
without risk of losing data.